Makes test script more resilient to aborts (#118)

+++ tests.ps1
@@ -32,7 +32,6 @@ $FirebirdConfiguration = @{
 $testsBaseDir = "$baseDir\src\FirebirdSql.Data.FirebirdClient.Tests"
 $testsProviderDir = "$testsBaseDir\bin\$Configuration\$(Get-UsedTargetFramework)"
 
-$startDir = $null
 $firebirdProcess = $null
 
 if ($env:tests_firebird_dir) {
@@ -45,7 +44,6 @@ else {
 function Prepare() {
 	echo "=== $($MyInvocation.MyCommand.Name) ==="
 
-	$script:startDir = $pwd
 	$selectedConfiguration = $FirebirdConfiguration[$FirebirdSelection]
 	$fbDownload = $selectedConfiguration.Download
 	$fbDownloadName = $fbDownload -Replace '.+/(.+)$','$1'
@@ -53,36 +51,46 @@ function Prepare() {
 		rm -Force -Recurse $firebirdDir
 	}
 	mkdir $firebirdDir | Out-Null
-	cd $firebirdDir
-	echo "Downloading $fbDownload"
-	Invoke-RestMethod -Uri $fbDownload -OutFile $fbDownloadName 
-	echo "Extracting $fbDownloadName"
-	7z x -bsp0 -bso0 $fbDownloadName
-	rm $fbDownloadName
-	cp -Recurse -Force .\* $testsProviderDir
 
-	ni firebird.log -ItemType File | Out-Null
+	pushd $firebirdDir
+	try {
+		echo "Downloading $fbDownload"
+		Invoke-RestMethod -Uri $fbDownload -OutFile $fbDownloadName 
+		echo "Extracting $fbDownloadName"
+		7z x -bsp0 -bso0 $fbDownloadName
+		rm $fbDownloadName
+		cp -Recurse -Force .\* $testsProviderDir
 
-	echo "Starting Firebird"
-	$process = Start-Process -FilePath $selectedConfiguration.Executable -ArgumentList $selectedConfiguration.Args -PassThru
-	echo "Version: $($process.MainModule.FileVersionInfo.FileVersion)"
-	$script:firebirdProcess = $process
+		ni firebird.log -ItemType File | Out-Null
 
-	echo "=== END ==="
+		echo "Starting Firebird"
+		$process = Start-Process -FilePath $selectedConfiguration.Executable -ArgumentList $selectedConfiguration.Args -PassThru
+		echo "Version: $($process.MainModule.FileVersionInfo.FileVersion)"
+		$script:firebirdProcess = $process
+
+		echo "=== END ==="
+	}
+	finally {
+		popd
+	}
 }
 
 function Cleanup() {
-	echo "=== $($MyInvocation.MyCommand.Name) ==="
+	# Do not write to output (Write-Output, echo) here. Write-Host is ok.
+	# -- https://stackoverflow.com/a/45105609
 
-	cd $script:startDir
+	Write-Host "=== $($MyInvocation.MyCommand.Name) ==="
+
 	$process = $script:firebirdProcess
-	$process.Kill()
-	$process.WaitForExit()
-	# give OS time to release all files
-	sleep -Milliseconds 100
+	if ($process) {
+		$process.Kill()
+		$process.WaitForExit()
+		# give OS time to release all files
+		sleep -Milliseconds 100
+	}
 	rm -Force -Recurse $firebirdDir
 
-	echo "=== END ==="
+	Write-Host "=== END ==="
 }
 
 function Tests-All() {
@@ -112,26 +120,46 @@ function Tests-FirebirdClient-Embedded() {
 	Tests-FirebirdClient 'Embedded' $False 'Disabled'
 }
 function Tests-FirebirdClient($serverType, $compression, $wireCrypt) {
-	cd $testsProviderDir
-	.\FirebirdSql.Data.FirebirdClient.Tests.exe --labels=All "--where=(ServerType==$serverType && Compression==$compression && WireCrypt==$wireCrypt) || Category==NoServer"
-	Check-ExitCode
+	pushd $testsProviderDir
+	try {
+		.\FirebirdSql.Data.FirebirdClient.Tests.exe --labels=All "--where=(ServerType==$serverType && Compression==$compression && WireCrypt==$wireCrypt) || Category==NoServer"
+		Check-ExitCode
+	}
+	finally {
+		popd
+	}
 }
 
 function Tests-EF6() {
-	cd "$baseDir\src\EntityFramework.Firebird.Tests\bin\$Configuration\$(Get-UsedTargetFramework)"
-	.\EntityFramework.Firebird.Tests.exe --labels=All
-	Check-ExitCode
+	pushd "$baseDir\src\EntityFramework.Firebird.Tests\bin\$Configuration\$(Get-UsedTargetFramework)"
+	try {
+		.\EntityFramework.Firebird.Tests.exe --labels=All
+		Check-ExitCode
+	}
+	finally {
+		popd
+	}
 }
 
 function Tests-EFCore() {
-	cd "$baseDir\src\FirebirdSql.EntityFrameworkCore.Firebird.Tests\bin\$Configuration\$(Get-UsedTargetFramework)"
-	.\FirebirdSql.EntityFrameworkCore.Firebird.Tests.exe --labels=All
-	Check-ExitCode
+	pushd "$baseDir\src\FirebirdSql.EntityFrameworkCore.Firebird.Tests\bin\$Configuration\$(Get-UsedTargetFramework)"
+	try {
+		.\FirebirdSql.EntityFrameworkCore.Firebird.Tests.exe --labels=All
+		Check-ExitCode
+	}
+	finally {
+		popd
+	}
 }
 function Tests-EFCore-Functional() {
-	cd "$baseDir\src\FirebirdSql.EntityFrameworkCore.Firebird.FunctionalTests"
-	dotnet test --no-build -c $Configuration
-	Check-ExitCode
+	pushd "$baseDir\src\FirebirdSql.EntityFrameworkCore.Firebird.FunctionalTests"
+	try {
+		dotnet test --no-build -c $Configuration
+		Check-ExitCode
+	}
+	finally {
+		popd
+	}
 }
 
 try {
